Refactored the structure. Moved the data and domain layer to a individual modules
