+++ .idea/androidTestResultsUserPreferences.xml
@@ -81,6 +81,7 @@
               <option name="preferredColumnWidths">
                 <map>
                   <entry key="Duration" value="172" />
+                  <entry key="Pixel_6_Pro_API_33" value="120" />
                   <entry key="Tests" value="441" />
                   <entry key="adb-JFDURKFAEMDIX8KR-HGkzTD._adb-tls-connect._tcp." value="120" />
                   <entry key="realme&#10; RMX3393" value="201" />
