--- app/src/main/java/com/nightcoder/vehiclesearch/data/state/ApiResult.kt
@@ -1,6 +1,0 @@
-package com.nightcoder.vehiclesearch.data.state
-
-sealed class ApiResult<out T> {
-    data class Success<out T>(val data: T) : ApiResult<T>()
-    data class Failed(val error: String) : ApiResult<Nothing>()
-}
