--- app/src/main/java/com/nightcoder/vehiclesearch/domain/model/VehicleFeatureItem.kt
@@ -1,7 +1,0 @@
-package com.nightcoder.vehiclesearch.domain.model
-
-data class VehicleFeatureItem(
-    val title: String,
-    val value: String,
-    val isHighlighted: Boolean = false
-)
