--- app/src/main/java/com/nightcoder/vehiclesearch/domain/model/VehicleInfoModel.kt
@@ -1,12 +1,0 @@
-package com.nightcoder.vehiclesearch.domain.model
-
-data class VehicleInfoModel(
-    val make: String,
-    val model: String,
-    val details: String,
-    val engine: String,
-    val gearbox: String,
-    val bodyType: String,
-    val year: String,
-    val motExpiry: String
-)
