--- app/src/main/java/com/nightcoder/vehiclesearch/logger/FileLogger.kt
@@ -1,13 +1,0 @@
-package com.nightcoder.vehiclesearch.logger
-
-import java.lang.Exception
-
-class FileLogger :Logger {
-    override fun info(msg: String) {
-
-    }
-
-    override fun error(message: String?, exception: Exception?) {
-
-    }
-}
