--- app/src/main/java/com/nightcoder/vehiclesearch/logger/Logger.kt
@@ -1,7 +1,0 @@
-package com.nightcoder.vehiclesearch.logger
-
-interface Logger {
-    fun info(msg: String): Unit
-    fun error(message:String?,exception: java.lang.Exception?)
-
-}
