+++ app/src/main/java/com/nightcoder/vehiclesearch/ui/features/searchtool/SearchToolActivity.kt
@@ -11,7 +11,7 @@ import androidx.core.view.isVisible
 import com.nightcoder.vehiclesearch.R
 import com.nightcoder.vehiclesearch.ui.adapter.VehicleFeatureGridAdapter
 import com.nightcoder.vehiclesearch.databinding.ActivitySearchToolBinding
-import com.nightcoder.vehiclesearch.domain.state.UiState
+import com.vehicleapp.domain.state.UiState
 import dagger.hilt.android.AndroidEntryPoint
