+++ app/src/main/java/com/nightcoder/vehiclesearch/ui/features/searchtool/SearchToolViewModel.kt
@@ -5,13 +5,13 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.nightcoder.vehiclesearch.data.state.ApiResult
+import com.vehicleapp.domain.state.ApiResult
 import com.nightcoder.vehiclesearch.di.IoDispatcher
-import com.nightcoder.vehiclesearch.domain.model.VehicleFeatureItem
-import com.nightcoder.vehiclesearch.domain.model.VehicleInfoModel
-import com.nightcoder.vehiclesearch.domain.state.UiState
-import com.nightcoder.vehiclesearch.domain.usecases.SearchVehicleInfoUseCase
-import com.nightcoder.vehiclesearch.logger.Logger
+import com.vehicleapp.domain.model.VehicleFeatureItem
+import com.vehicleapp.domain.model.VehicleInfoModel
+import com.vehicleapp.domain.state.UiState
+import com.vehicleapp.domain.usecases.SearchVehicleInfoUseCase
+import com.vehicleapp.domain.logger.Logger
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.launch
