+++ domain/src/main/java/com/vehicleapp/domain/logger/FileLogger.kt
@@ -1,0 +1,13 @@
+package com.vehicleapp.domain.logger
+
+import java.lang.Exception
+
+class FileLogger : Logger {
+    override fun info(msg: String) {
+
+    }
+
+    override fun error(message: String?, exception: Exception?) {
+
+    }
+}
