+++ domain/src/main/java/com/vehicleapp/domain/logger/Logger.kt
@@ -1,0 +1,7 @@
+package com.vehicleapp.domain.logger
+
+interface Logger {
+    fun info(msg: String): Unit
+    fun error(message:String?,exception: java.lang.Exception?)
+
+}
