+++ domain/src/main/java/com/vehicleapp/domain/model/VehicleFeatureItem.kt
@@ -1,0 +1,7 @@
+package com.vehicleapp.domain.model
+
+data class VehicleFeatureItem(
+    val title: String,
+    val value: String,
+    val isHighlighted: Boolean = false
+)
