+++ domain/src/main/java/com/vehicleapp/domain/model/VehicleInfoModel.kt
@@ -1,0 +1,12 @@
+package com.vehicleapp.domain.model
+
+data class VehicleInfoModel(
+    val make: String,
+    val model: String,
+    val details: String,
+    val engine: String,
+    val gearbox: String,
+    val bodyType: String,
+    val year: String,
+    val motExpiry: String
+)
