+++ domain/src/main/java/com/vehicleapp/domain/service/SearchVehicleService.kt
@@ -1,0 +1,11 @@
+package com.vehicleapp.domain.service
+
+import com.vehicleapp.domain.model.VehicleInfoModel
+import com.vehicleapp.domain.state.ApiResult
+
+
+interface SearchVehicleService {
+   suspend fun queryVehicleInfo(regNo: String): ApiResult<VehicleInfoModel>
+}
+
+
